+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# v 1.5.0 (2022-04-29)
+Changes in this release:
+- Add support for iso5 container environment (#192)
+- Add support for SSL and anthentication (#186)
+- Report skipped and deploying resources when reaching a bad lifecycle state (#199)
+- Fix the iso4 jenkins job by adding a constraint on the lsm module's version
+- Add the possibility to add other constraints through the INMANTA_LSM_MODULE_CONSTRAINTS environment variable
+
 # v 1.4.1 (2022-02-10)
 Changes in this release:
 - Run project installation with server's environment variables